requirements.txt
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.2.2
dvc==3.30.1
PyYAML==6.0.1
matplotlib==3.7.1
seaborn==0.12.2